==> 01_extract_coli.ipynb <==
# import modules
import pandas as pd

## Cost of Living Index Data

# parse in html tables and check for number of records per table
years = [2015, 2016, 2017, 2018, 2019]

all_COLI_data = []

for year in years:
    url = 'https://www.numbeo.com/cost-of-living/rankings_by_country.jsp?title=' + str(year)
    print(url)
    COLI_data = pd.read_html(url)
    print(len(COLI_data[2]))
    all_COLI_data.append(COLI_data[2])

# save COLI data as csv in 01_extract_coli
for i in range(5):
    outpath="./01_extract_coli/"+str(years[i])+"coli.csv"
    all_COLI_data[i].to_csv(outpath,index=False)

## Property Index Data

# parse in html tables and check for number of records per table
all_property_data = []

for year in years:
    url = 'https://www.numbeo.com/property-investment/rankings_by_country.jsp?title=' + str(year)
    print(url)
    property_data = pd.read_html(url)
    print(len(property_data[2]))
    all_property_data.append(property_data[2])

# save property data as csv in 01_extract_coli
for i in range(5):
    outpath="./01_extract_coli/"+str(years[i])+"property.csv"
    all_property_data[i].to_csv(outpath,index=False)
