==> 01_extract_country.ipynb <==
# import dependencies
from splinter import Browser
from webdriver_manager.chrome import ChromeDriverManager
from bs4 import BeautifulSoup
from time import sleep

# Mac User run this cell and Ignore cells 3  
executable_path={'executable_path':'/usr/local/bin/chromedriver'}

# Windows User Please Run this Cell 
executable_path = {'executable_path': ChromeDriverManager().install()}

#Setup instance of Browser
browser=Browser('chrome',**executable_path,headless=False)

# path to country and continent list
url='https://datahub.io/JohnSnowLabs/country-and-continent-codes-list'
browser.visit(url)
sleep(1)
browser.links.find_by_partial_text('original version').click()

# use Beautiful Soup to read text
html=browser.html
soup=BeautifulSoup(html,'lxml')
result=soup.body.text
print(result)

# save to 01_extract_country folder
textfile = open('01_extract_country/country.csv', 'w')
textfile.write(result)
textfile.close()

browser.quit()
