==> 01_extract_happiness.ipynb <==
# to use kaggle api command line, need
# pip install kaggle
# need kaggle api key
# see https://github.com/Kaggle/kaggle-api
import sys
sys.path.append('./00_config')
from api_key_template import username,key
import os

# set kaggle username and api key
os.environ['KAGGLE_USERNAME'] = username
os.environ['KAGGLE_KEY'] = key

# authenticate api key
# ref https://technowhisp.com/kaggle-api-python-documentation/
from kaggle.api.kaggle_api_extended import KaggleApi
api = KaggleApi()
api.authenticate()

# commands based off kaggle_api_extended.py usage
# https://github.com/Kaggle/kaggle-api/blob/master/kaggle/api/kaggle_api_extended.py
# viewing config
api.print_config_values()

# set download file path
api.set_config_value('path','./01_extract_happiness/')

api.print_config_values()

# list world happiness file
api.dataset_list_files('unsdsn/world-happiness').files

# download files
api.dataset_download_files('unsdsn/world-happiness',unzip=True)
