==> 02_transform_coli.ipynb <==
# import modules
import os
import pandas as pd
import numpy as np

## Cost Of Living Index Data

# create a dataframe for each year
COLI_df_2015 = pd.read_csv('01_extract_coli/2015coli.csv')
COLI_df_2016 = pd.read_csv('01_extract_coli/2016coli.csv')
COLI_df_2017 = pd.read_csv('01_extract_coli/2017coli.csv')
COLI_df_2018 = pd.read_csv('01_extract_coli/2018coli.csv')
COLI_df_2019 = pd.read_csv('01_extract_coli/2019coli.csv')

# add a column for year and assigning year value
COLI_df_2015["year"] = int(2015)
COLI_df_2016["year"] = int(2016)
COLI_df_2017["year"] = int(2017)
COLI_df_2018["year"] = int(2018)
COLI_df_2019["year"] = int(2019) 

# add annual ranking based on row number + 1
COLI_df_2015["ranking"] = np.arange(start = 1, stop = len(COLI_df_2015)+1, step = 1)
COLI_df_2016["ranking"] = np.arange(start = 1, stop = len(COLI_df_2016)+1, step = 1)
COLI_df_2017["ranking"] = np.arange(start = 1, stop = len(COLI_df_2017)+1, step = 1)
COLI_df_2018["ranking"] = np.arange(start = 1, stop = len(COLI_df_2018)+1, step = 1)
COLI_df_2019["ranking"] = np.arange(start = 1, stop = len(COLI_df_2019)+1, step = 1)

# drop rank column
COLI_df_2015 = COLI_df_2015.drop(["Rank"], axis=1)
COLI_df_2016 = COLI_df_2016.drop(["Rank"], axis=1)
COLI_df_2017 = COLI_df_2017.drop(["Rank"], axis=1)
COLI_df_2018 = COLI_df_2018.drop(["Rank"], axis=1)
COLI_df_2019 = COLI_df_2019.drop(["Rank"], axis=1)

# concatenate dataframes for all years into a single dataframe
COLI_result_df = pd.concat([COLI_df_2015, COLI_df_2016, COLI_df_2017, 
                                COLI_df_2018, COLI_df_2019
                               ], 
                               axis=0, 
                               join="inner"
                              )
COLI_result_df=COLI_result_df.reset_index(drop=True)
COLI_result_df

# rename columns to remove spacing (required for SQL database)
COLI_result_df.rename(columns={"Country":"country",
                        "Cost of Living Index":"costofliving_index",
                        "Rent Index":"rent_index",
                        "Groceries Index":"groceries_index",
                        "Restaurant Price Index":"restaurantprice_index",
                        "Local Purchasing Power Index":"localpurchpower_index"
                       }, 
               inplace=True
              )

# drop redundant columns
COLI_result_df.drop(["Cost of Living Plus Rent Index"], axis=1)

# resize dataframe (keeping relevant columns) and rearrange column order
COLI_result_df = pd.DataFrame(COLI_result_df[["country", "year", "ranking", "costofliving_index", "rent_index", "groceries_index", 
                   "restaurantprice_index","localpurchpower_index"
                  ]
                 ]
                             ) 
COLI_result_df


# count the number of years data for each country
country_df = pd.DataFrame(COLI_result_df.groupby(["country"])["year"].count())

# iterate over dataframe to drop countries where there are not 5 years of data
for index, row in country_df.iterrows():
    if row["year"] !=5:
        print(f"{index} dropped")
        country_df.drop(index, inplace = True)

# validate that all countries without 5 years of data have been dropped
check = country_df.loc[country_df["year"] != 5]
check

# filter the database for countries with only 5 years of data
countries = list(country_df.index.values) 
countries

# filter datafrome for countries with 5 years of data
COLI_df = pd.DataFrame(COLI_result_df[COLI_result_df["country"].isin(countries)])
COLI_df

# review simple stats as a reasonableness check (the "sniff" test)
COLI_df.describe()

# confirm same number of records per year for all years
check_df = pd.DataFrame(COLI_df.groupby(["year"]).count())
check_df

# import country&continent csv
country=os.path.join('02_transform_country','country.csv')
country_df=pd.read_csv(country)
country_df.head()

# replace countries with country code
for ind,row in COLI_df.iterrows():
    for ind2,row2 in country_df.iterrows():
        if row['country'].lower() in row2['country'].lower():
            COLI_df.loc[ind,'country']=country_df.loc[ind2,'countrycode']
            break

# find countries without a country code from the main dataframe
miscountry=[]
for i in COLI_df['country']:
    if (len(i)>3) and i not in miscountry:
        miscountry.append(i)
print(miscountry)

# invert match to capture countries that are named differently
# ie. previous round looking for happyall country string within continent_df country, 
# now looking for continent_df country string within happyall country
for ind,row in country_df.iterrows():
    for j in miscountry:
        if row['country'].lower() in j.lower():
            COLI_df['country']=COLI_df['country'].replace(j,country_df.loc[ind,'countrycode'])

# find countries without a country code from the main dataframe
miscountry=[]
for i in COLI_df['country']:
    if (len(i)>3) and i not in miscountry:
        miscountry.append(i)
print(miscountry)

# rename manually but only if country code exist in country list
countryCode=list(country_df['countrycode'])
if 'PSE' in countryCode:
    COLI_df['country'] = COLI_df['country'].replace(['Palestine'],['PSE'])
else:
    COLI_df=COLI_df[COLI_df.country!='Palestine']

# ensure that all countries still have 5 years of data
countrycount=pd.DataFrame(COLI_df.groupby('country')['year'].count())
for ind,row in countrycount.iterrows():
    if row['year']!=5:
        print(f'{ind}-{row["year"]}')

COLI_df=COLI_df.rename(columns={'country':'countrycode'})

# save transformed cost of living index data
COLI_df.to_csv('02_transform_coli/coli_data.csv',index=False)

## Property Index Data

# create a dataframe for each year
property_df_2015 = pd.read_csv('01_extract_coli/2015property.csv')
property_df_2016 = pd.read_csv('01_extract_coli/2016property.csv')
property_df_2017 = pd.read_csv('01_extract_coli/2017property.csv')
property_df_2018 = pd.read_csv('01_extract_coli/2018property.csv')
property_df_2019 = pd.read_csv('01_extract_coli/2019property.csv')

# add a column for year and assigning year value
property_df_2015["Year"] = int(2015)
property_df_2016["Year"] = int(2016)
property_df_2017["Year"] = int(2017)
property_df_2018["Year"] = int(2018)
property_df_2019["Year"] = int(2019)

# drop rank column
property_df_2015 = property_df_2015.drop(["Rank"], axis=1)
property_df_2016 = property_df_2016.drop(["Rank"], axis=1)
property_df_2017 = property_df_2017.drop(["Rank"], axis=1)
property_df_2018 = property_df_2018.drop(["Rank"], axis=1)
property_df_2019 = property_df_2019.drop(["Rank"], axis=1)


# concatenate dataframes for all years to find common countries across all years
property_result_df = pd.concat([property_df_2015,property_df_2016, property_df_2017, 
                                property_df_2018, property_df_2019
                               ], 
                               axis=0, 
                               join="inner"
                              )
property_result_df=property_result_df.reset_index(drop=True)
property_result_df

# resize dataframe (keeping relevant columns) and rearrange column order
property_result_df = pd.DataFrame(property_result_df[["Country", "Year", "Price To Income Ratio", "Mortgage As A Percentage Of Income", 
                           "Affordability Index"
                          ]
                         ]
                          )
property_result_df

# rename columns to remove spacing (required for SQL database)
property_result_df.rename(columns={"Country":"country",
                            "Year":"year",
                            "Price To Income Ratio":"price_income_ratio",
                            "Mortgage As A Percentage Of Income":"mortage_income_ratio",
                            "Affordability Index":"affordability_index"
                           }, 
                   inplace=True
                  )
property_result_df

# count the number of years data for each country
pcountry_df = pd.DataFrame(property_result_df.groupby(["country"])["year"].count())

# iterate over dataframe to drop countries where there are not 5 years of data
for index, row in pcountry_df.iterrows():
    if row["year"] !=5:
        pcountry_df.drop(index, inplace = True)
        
pcountry_df

# validate that all countries without 5 years of data have been dropped
check = pcountry_df.loc[pcountry_df["year"] != 5]
check

# filter the database for countries with only 5 years of data
final_countries = list(pcountry_df.index.values) 
final_countries

# filter property datafrome for countries with 5 years of data
property_df = pd.DataFrame(property_result_df[property_result_df["country"].isin(final_countries)])
property_df

# review simple stats as a reasonableness check (the "sniff" test)
property_df.describe()

# confirm same number of records per year for all years
check_df = pd.DataFrame(property_df.groupby(["year"]).count())
check_df

# replace countries with country code
for ind,row in property_df.iterrows():
    for ind2,row2 in country_df.iterrows():
        if row['country'].lower() in row2['country'].lower():
            property_df.loc[ind,'country']=country_df.loc[ind2,'countrycode']
            break

# find countries without a country code from the main dataframe
miscountry=[]
for i in property_df['country']:
    if (len(i)>3) and i not in miscountry:
        miscountry.append(i)
print(miscountry)

# invert match to capture countries that are named differently
# ie. previous round looking for happyall country string within continent_df country, 
# now looking for continent_df country string within happyall country
for ind,row in country_df.iterrows():
    for j in miscountry:
        if row['country'].lower() in j.lower():
            property_df['country']=property_df['country'].replace(j,country_df.loc[ind,'countrycode'])

# find countries without a country code from the main dataframe
miscountry=[]
for i in property_df['country']:
    if (len(i)>3) and i not in miscountry:
        miscountry.append(i)
print(miscountry)

# rename manually but only if country code exist in country list
countryCode=list(country_df['countrycode'])
if 'PSE' in countryCode:
    property_df['country'] = property_df['country'].replace(['Palestine'],['PSE'])
else:
    property_df=property_df[property_df.country!='Palestine']

# ensure that all countries still have 5 years of data
countrycount=pd.DataFrame(property_df.groupby('country')['year'].count())
for ind,row in countrycount.iterrows():
    if row['year']!=5:
        print(f'{ind}-{row["year"]}')

property_df=property_df.rename(columns={'country':'countrycode'})

# save transformed property index data
property_df.to_csv('02_transform_coli/property_data.csv',index=False)
