==> 02_transform_country.ipynb <==
# import modules
import os
import pandas as pd

# read country csv
path=os.path.join('01_extract_country','country.csv')
country_df=pd.read_csv(path)
country_df.head()

# clean country name
for ind,row in country_df.iterrows():
    country_df.loc[ind,'Country_Name']=country_df.loc[ind,'Country_Name'].split(',')[0]
    country_df.loc[ind,'Country_Name']=country_df.loc[ind,'Country_Name'].split('(')[0]
    country_df.loc[ind,'Country_Name']=country_df.loc[ind,'Country_Name'].strip()
country_df.head()

# drop duplicate country code and country
country_df=country_df.drop_duplicates(subset =["Three_Letter_Country_Code"], keep = 'last')
country_df=country_df.drop_duplicates(subset =["Country_Name"], keep = 'last')

# find problem names
probcountry=[]
for country in country_df['Country_Name']:
    for string in country.split(" "):   
        if not string.isalnum():
            probcountry.append(country)

print(probcountry)
print(len(probcountry))

# manually replace problem names (not considering "'","&",or "-" problematic)
# and verify that replacement has occured
country_df=country_df.replace([probcountry[0]],['Aland Islands'])

probcountry=[]
for country in country_df['Country_Name']:
    for string in country.split(" "):   
        if not string.isalnum():
            probcountry.append(country)

print(probcountry)
print(len(probcountry))

# note also that 'united states' has multiple entries which are confusing
# manually editing it
probcountry=[]
for country in country_df['Country_Name']:
    if 'united states' in country.lower():
        probcountry.append(country)
print(probcountry)

country_df=country_df[country_df['Country_Name']!=probcountry[0]]
country_df=country_df.replace([probcountry[2]],['Virgin Islands'])

# rename and rearrange columns
country_df_new=country_df[['Three_Letter_Country_Code','Country_Name','Continent_Name']]
country_df_new.columns=['countrycode','country','continent']
country_df_new.head()

# drop nulls
country_df_new=country_df_new.dropna(axis=0)

# save country csv for SQL database
country_df_new.to_csv('02_transform_country/country.csv',index=False)
