==> 02_transform_happiness.ipynb <==
# import modules
import os
import pandas as pd
import numpy as np

# import happiness report csv files fo 2015-2019 
year_list = ['2015', '2016', '2017', '2018', '2019']
path =['./01_extract_happiness/datasets/unsdsn/world-happiness/{Year}.csv'.format(Year = year) for year in year_list]

# formatting dataset
def formatDt(data, year):
    data.insert(loc=0, column='Year', value=[year]*len(data))
    return data

data={}
for i in range(len(year_list)):
    data[year_list[i]]=formatDt(pd.read_csv(path[i]), year_list[i])
    print(list(data[year_list[i]].columns))

data['2018'].columns = ['Year', 'Overall rank', 'Country or region', 'Score', 'GDP per capita', 'Family', 'Healthy life expectancy', 'Freedom to make life choices', 'Generosity', 'Perceptions of corruption']
data['2019'].columns = ['Year', 'Overall rank', 'Country or region', 'Score', 'GDP per capita', 'Family', 'Healthy life expectancy', 'Freedom to make life choices', 'Generosity', 'Perceptions of corruption']

for i in range(len(year_list)):
    print(list(data[year_list[i]].columns))

#drop columns
def intersection(lst1, lst2): 
    lst3 = [value for value in lst1 if value not in lst2]
    lst4 = [value for value in lst2 if value not in lst1]
    print(lst3)
    print(lst4)

intersection(list(data['2015'].columns), list(data['2016'].columns))

data['2015'].drop('Standard Error', axis=1, inplace=True)
data['2016'].drop(['Lower Confidence Interval','Upper Confidence Interval'], axis=1, inplace=True)
data['2015']

intersection(list(data['2016'].columns), list(data['2017'].columns))

data['2016'].drop('Region', axis=1, inplace=True)
data['2015'].drop('Region', axis=1, inplace=True)
data['2017'].drop(['Whisker.high', 'Whisker.low'], axis=1, inplace=True)

intersection(list(data['2017'].columns), list(data['2018'].columns))

columns_titles = ['Year', 'Country or region', 'Overall rank', 'Score', 'GDP per capita', 'Family', 'Healthy life expectancy', 'Freedom to make life choices', 'Generosity', 'Perceptions of corruption']
data['2018']=data['2018'].reindex(columns=columns_titles)
data['2018']

intersection(list(data['2017'].columns), list(data['2018'].columns))

data['2017'].drop(['Dystopia.Residual'], axis=1, inplace=True)

intersection(list(data['2018'].columns), list(data['2019'].columns))

columns_titles = ['Year', 'Country or region', 'Overall rank', 'Score', 'GDP per capita', 'Family','Healthy life expectancy', 'Freedom to make life choices', 'Generosity', 'Perceptions of corruption']
data['2019']=data['2019'].reindex(columns=columns_titles)
data['2019']

for i in range(len(year_list)):
    print(list(data[year_list[i]].columns))

for i in range(len(year_list)):
    if 'Dystopia Residual' in (list(data[year_list[i]].columns)):
        data[year_list[i]].drop(['Dystopia Residual'], axis=1, inplace=True)

for i in range(len(year_list)):
    print(list(data[year_list[i]].columns))

for i in range(len(year_list)):
    print(len(data[year_list[i]].columns))

columns_titles = ['year', 'country', 'happyrank', 'happyscore', 'economy', 'social','health', 'freedom', 'generosity', 'trust']
data['2015'].columns = columns_titles
data['2016'].columns = columns_titles

columns_titles = ['year', 'country', 'happyrank', 'happyscore', 'economy', 'social', 'health', 'freedom', 'generosity', 'trust']
data['2017'].columns = columns_titles
data['2018'].columns = columns_titles
data['2019'].columns = columns_titles

for i in range(len(year_list)):
    print(list(data[year_list[i]].columns))

data['2015']

# concat all 5 years worth of happiness data into a single df
happyall=pd.concat([data['2015'],data['2016'],data['2017'],data['2018'],data['2019']],axis=0)
happyall=happyall.reset_index(drop=True)
happyall

# find countries without all five years of data
countrycount=pd.DataFrame(happyall.groupby('country')['year'].count())
for ind,row in countrycount.iterrows():
    if row['year']!=5:
        print(f'{ind}-{row["year"]}')

# replace problematic country names
happyall['country'] = happyall['country'].replace(\
    ['Hong Kong S.A.R., China','Northern Cyprus', 'Somaliland region','Taiwan Province of China','Trinidad & Tobago'],\
    ['Hong Kong','North Cyprus','Somaliland Region','Taiwan','Trinidad and Tobago'])

# drop countries without all five years of data
countrycount=pd.DataFrame(happyall.groupby('country')['year'].count())
countrycount
for ind,row in countrycount.iterrows():
    if row['year']!=5:
        happyall=happyall[happyall.country!=ind]

happyall

# import country&continent csv
country=os.path.join('02_transform_country','country.csv')
country_df=pd.read_csv(country)

# replace countries with country code
for ind,row in happyall.iterrows():
    for ind2,row2 in country_df.iterrows():
        if row['country'].lower() in row2['country'].lower():
            happyall.loc[ind,'country']=country_df.loc[ind2,'countrycode']
            break

# find countries without a country code from the main dataframe
miscountry=[]
for i in happyall['country']:
    if (len(i)>3) and i not in miscountry:
        miscountry.append(i)
print(miscountry)

# invert match to capture countries that are named differently
# ie. previous round looking for happyall country string within continent_df country, 
# now looking for continent_df country string within happyall country
# takes care of situation such as happyall country of 'Congo (Brazzaville)' not in the continent_df country of 'Congo'
for ind,row in country_df.iterrows():
    for j in miscountry:
        if row['country'].lower() in j.lower():
            happyall['country']=happyall['country'].replace(j,country_df.loc[ind,'countrycode'])

# find all missed countries again
miscountry=[]
for i in happyall['country']:
    if (len(i)>3) and i not in miscountry:
        miscountry.append(i)
print(miscountry)

# last ditch attempt to get final few countries' alternate name by scrapping wiki
from splinter import Browser
from webdriver_manager.chrome import ChromeDriverManager
from bs4 import BeautifulSoup
from time import sleep
executable_path={'executable_path':'/usr/local/bin/chromedriver'}
browser=Browser('chrome',**executable_path,headless=False)
altname=[None]*len(miscountry)
for i in range(len(miscountry)):
    url='http://wikipedia.org'
    browser.visit(url)
    sleep(1)
    browser.find_by_id('searchInput').fill(f'{miscountry[i]}\n')
    html=browser.html
    soup=BeautifulSoup(html,'lxml')
    result=soup.find_all('div', class_='country-name')
    for r in result:
        altname[i]=r.text
browser.quit()

print(altname)

for ind,row in country_df.iterrows():
    for j in range(len(miscountry)):
        if row['country'].lower().split(" ")[0] in altname[j].lower():
            happyall['country']=happyall['country'].replace(miscountry[j],country_df.loc[ind,'countrycode'])

# still missed a few countries
miscountry=[]
for i in happyall['country']:
    if (len(i)>3) and i not in miscountry:
        miscountry.append(i)
print(miscountry)

# brute forcing it
happyall['country'] = happyall['country'].replace(\
    ['Kosovo','Ivory Coast'],\
    ['XKX','CIV'])

# find countries without five years of data
countrycount=pd.DataFrame(happyall.groupby('country')['year'].count())
countrycount
for ind,row in countrycount.iterrows():
    if row['year']!=5:
        print(f'{ind}-{row["year"]}')

# problem resulted from there being both 'republic of congo' and 'democratic republic of congo'; 
# as well as 'cyprus' and 'northern cyprus' having the same 3-digit code
# dropping both COG and CYP
happyall=happyall[happyall.country!='COG']
happyall=happyall[happyall.country!='CYP']

# drop country not in country list
countryCode=list(country_df['countrycode'])
for ind,row in happyall.iterrows():
    if row['country'] not in countryCode:
        happyall=happyall.drop(ind)
        print(f"dropped country {row['country']}")

# number of unique countries
len(countrycount)

# check for missing values
happyall.count()

# remove country with missing trust value
for ind,row in happyall.iterrows():
    try:
        int(row['trust'])
    except:
        country=row['country']
        print(row['country'])

happyall=happyall[happyall.country!='ARE']
happyall.count()

happyall.describe()

happyall=happyall.rename(columns={'country':'countrycode'})

# save transformed happiness data
happyall.to_csv('02_transform_happiness/happy_data.csv',index=False)
