==> 03_load.ipynb <==
# import dependencies
import os
import pandas as pd
import sys
sys.path.append('./00_config')
from password_template import username, password
from sqlalchemy import create_engine, inspect
from sqlalchemy_utils import database_exists, create_database

# create engine and database
engine = create_engine('postgresql://'+username+':'+password+'@localhost/country_happyCOL')
if not database_exists(engine.url):
    create_database(engine.url)

# set csv import path
countrypath=os.path.join('02_transform_country','country.csv')
happypath=os.path.join('02_transform_happiness','happy_data.csv')
colipath=os.path.join('02_transform_coli','coli_data.csv')
propertypath=os.path.join('02_transform_coli','property_data.csv')

# import country data
countrydata=pd.read_csv(countrypath)
countrydata.head()

# create year table
yeardata=pd.DataFrame({"year":["2015","2016","2017","2018","2019"]})
yeardata

# import happiness data
happydata=pd.read_csv(happypath)
happydata.head()

# import COLI data
colidata=pd.read_csv(colipath)
colidata.head()

# import property data
propertydata=pd.read_csv(propertypath)
propertydata.head()

#drop table if exists
engine.execute("DROP TABLE IF EXISTS coli")
engine.execute("DROP TABLE IF EXISTS property")
engine.execute("DROP TABLE IF EXISTS happiness")
engine.execute("DROP TABLE IF EXISTS country")
engine.execute("DROP TABLE IF EXISTS year")

# create country table & constraints in SQL
engine.execute("CREATE TABLE country (\
                    countrycode CHAR(3) NOT NULL PRIMARY KEY,\
                    country VARCHAR(100) NOT NULL,\
                    continent VARCHAR(100) NOT NULL)")

# save country data into SQL database
countrydata.to_sql(name='country', con=engine, if_exists='append', index=False)

# create year table & constraints in SQL
engine.execute("CREATE TABLE year (year INT NOT NULL PRIMARY KEY)")

# save year data into SQL database
yeardata.to_sql(name='year', con=engine, if_exists='append', index=False)

# create happiness table & constraints in SQL
engine.execute("CREATE TABLE happiness (\
                    id SERIAL PRIMARY KEY,\
                    countrycode CHAR(3) NOT NULL,\
                    year INT NOT NULL,\
                    happyrank INT NOT NULL,\
                    happyscore DECIMAL NOT NULL,\
                    economy DECIMAL NOT NULL,\
                    social DECIMAL NOT NULL,\
                    health DECIMAL NOT NULL,\
                    freedom DECIMAL NOT NULL,\
                    generosity DECIMAL NOT NULL,\
                    trust DECIMAL NOT NULL,\
                    FOREIGN KEY (countrycode)\
                    REFERENCES country(countrycode),\
                    FOREIGN KEY (year)\
                    REFERENCES year(year))")

# save happiness data into SQL database
happydata.to_sql(name='happiness', con=engine, if_exists='append', index=False)

# create COLI table & constraints in SQL
engine.execute("CREATE TABLE coli (\
                    id SERIAL PRIMARY KEY,\
                    countrycode CHAR(3) NOT NULL,\
                    year INT NOT NULL,\
                    ranking INT NOT NULL,\
                    costofliving_index DECIMAL NOT NULL,\
                    rent_index DECIMAL NOT NULL,\
                    groceries_index DECIMAL NOT NULL,\
                    restaurantprice_index DECIMAL NOT NULL,\
                    localpurchpower_index DECIMAL NOT NULL,\
                    FOREIGN KEY (countrycode)\
                    REFERENCES country(countrycode),\
                    FOREIGN KEY (year)\
                    REFERENCES year(year))")

# save COLI data into SQL database
colidata.to_sql(name='coli', con=engine, if_exists='append', index=False)

# create property table & constraints in SQL
engine.execute("CREATE TABLE property (\
                    id SERIAL PRIMARY KEY,\
                    countrycode CHAR(3) NOT NULL,\
                    year INT NOT NULL,\
                    price_income_ratio DECIMAL NOT NULL,\
                    mortage_income_ratio DECIMAL NOT NULL,\
                    affordability_index DECIMAL NOT NULL,\
                    FOREIGN KEY (countrycode)\
                    REFERENCES country(countrycode),\
                    FOREIGN KEY (year)\
                    REFERENCES year(year))")

# save property data into SQL database
propertydata.to_sql(name='property', con=engine, if_exists='append', index=False)

# ensure that tables are in database
inspector = inspect(engine)
tables=inspector.get_table_names()
print(tables)

# confirm each table exists in sql by querying it
for table in tables:
    query='select * from '+ table
    print(pd.read_sql_query(query, con=engine).head())
